refactor: restructure TimeSeriesGraph and introduce new components for enhanced visualization
- Refactored TimeSeriesGraph to improve data handling and visualization.
- Introduced new components: Area, Line, Points, XAxis, and YAxis for better modularity and clarity.
- Updated data fetching parameters in useList for optimized performance.
- Enhanced error handling for invalid data scenarios.
- Removed deprecated code and improved overall component structure.
